Update the right filter.js with chnages
Ability to add an order by toggle linking into the json query.

FJS.addOrder({ ele: '#idoforderby'});
expects the orderby as the value in option as jsonfield|direction e.g. price|asc

Also added support for multiple values in an input e.g.: <input value="1,2,3"

Lastly, added if no value pass a highly unlikely value to the filter. This will ensure that when nothing is sent to the filter instead of showing all, nothing is shown.