Update filter.js
Added the ability to use the like in json_query. This also allows for the AND search when adding filters.

Example:
FJS.addCriteria({field: 'feature', ele: '#featurewrap input:checkbox', type: 'like'});
<div id="featurewrap">
<input type="checkbox" name="feature" value="m_2403" >Value 1
<input type="checkbox" name="feature" value="m_2001" >Value 2
</div>

if the first is clicked value m_2403 must be in the feature property of the JSON. if the second is clicked as well then using regex both must be present.

Works only with strings